+++ data/network/src/MuseumApi.kt
@@ -1,28 +1,6 @@
 package com.jetbrains.kmpapp.data
 
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.request.get
-import io.ktor.utils.io.CancellationException
-
 interface MuseumApi {
     suspend fun getData(): List<MuseumObject>
 }
 
-class KtorMuseumApi(private val client: HttpClient) : MuseumApi {
-    companion object {
-        private const val API_URL =
-            "https://raw.githubusercontent.com/Kotlin/KMP-App-Template/main/list.json"
-    }
-
-    override suspend fun getData(): List<MuseumObject> {
-        return try {
-            client.get(API_URL).body()
-        } catch (e: Exception) {
-            if (e is CancellationException) throw e
-            e.printStackTrace()
-
-            emptyList()
-        }
-    }
-}
