+++ data/storage/src/InMemoryMuseumStorage.kt
@@ -1,0 +1,21 @@
+package com.jetbrains.kmpapp.data
+
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.map
+
+class InMemoryMuseumStorage : MuseumStorage {
+    private val storedObjects = MutableStateFlow(emptyList<MuseumObject>())
+
+    override suspend fun saveObjects(newObjects: List<MuseumObject>) {
+        storedObjects.value = newObjects
+    }
+
+    override fun getObjectById(objectId: Int): Flow<MuseumObject?> {
+        return storedObjects.map { objects ->
+            objects.find { it.objectID == objectId }
+        }
+    }
+
+    override fun getObjects(): Flow<List<MuseumObject>> = storedObjects
+}
