+++ data/storage/src/MuseumStorage.kt
@@ -1,8 +1,6 @@
 package com.jetbrains.kmpapp.data
 
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.map
 
 interface MuseumStorage {
     suspend fun saveObjects(newObjects: List<MuseumObject>)
@@ -12,18 +10,3 @@ interface MuseumStorage {
     fun getObjects(): Flow<List<MuseumObject>>
 }
 
-class InMemoryMuseumStorage : MuseumStorage {
-    private val storedObjects = MutableStateFlow(emptyList<MuseumObject>())
-
-    override suspend fun saveObjects(newObjects: List<MuseumObject>) {
-        storedObjects.value = newObjects
-    }
-
-    override fun getObjectById(objectId: Int): Flow<MuseumObject?> {
-        return storedObjects.map { objects ->
-            objects.find { it.objectID == objectId }
-        }
-    }
-
-    override fun getObjects(): Flow<List<MuseumObject>> = storedObjects
-}
